add frontend and controller

+++ src/test/java/com/MolvenoLakeResort/model/restaurant/TableTest.java
@@ -17,7 +17,7 @@ public class TableTest {
 
         this.table.setId(5);
 
-        int actual = this.table.getId();
+        long actual = this.table.getId();
         Assert.assertEquals(5, actual);
     }
